requirements.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8
django-filter>=24.0
psycopg2-binary>=2.9
dj-database-url>=2.1
